refactor(framework): extract the transaction cost model into TxGasModel.h
Moves the intrinsic-gas formula out of ethereum-executor into
bcos-framework/protocol/TxGasModel.h: `ethToAddress`, `num_words`,
`compute_tx_data_tokens`, `compute_access_list_cost`, `TransactionCost` and
`compute_tx_intrinsic_cost`, plus the MAX_TX_GAS_LIMIT and
AUTHORIZATION_EMPTY_ACCOUNT_COST constants.

Nothing changes behaviourally. ethereum-executor keeps every call site through
using-declarations, so `evm::MAX_TX_GAS_LIMIT` and
`eth_transition_detail::compute_tx_intrinsic_cost` still resolve.

The reason to move it is the admission layer being built on top of this: it has
to reject a transaction whose gasLimit cannot cover the intrinsic cost, and it
has to use this exact formula. Admission must never be STRICTER than execution,
and a second copy would drift at the first hard fork that moves EIP-7623's
min_cost or the authorization-list cost -- silently, and only for transactions
near the boundary.

`ethToAddress` moves with the rest because "is this a contract creation" is
defined as "the `to` field does not decode to an address". bcos::toAddress
(FixedBytes.h) cannot substitute: it throws on anything that is not exactly 20
bytes, and an empty `to` -- every deploy -- is the common case here, not an
error.

Verified: full build clean (0 errors); test-ethereum-state-smoke passes,
test-bcos-txpool and test-bcos-rpc report "No errors detected".
